Draft refactor to decouple command execution.
So we have a decoupling with the router and actions but the handler
functions still have many strings in the arguments list. So now
processing goes via the method assigned to the action definition which
starts to handle a matched route locating objects by string where needed
for the actual application logic it defers to command execute.

This allows primarily string based access for network interfaces and
native types when we want to invoke command functions directly.

It does mean that the following steps need to be followed to create a
new command:

1) Add a route action and command definition to the ActionList.
2) Declare the corresponding handler on CommandAdapter.
3) Proxy the CommandAdapter method to a CommandExecute method.

The convention is that the command adapter and command execute methods
should have exactly the same name although the arguments and return
values will differ.